moved workflow to base (See #507 comments)
